style(jobs): move swarm error to swarm job file
